README:Revisions.md
Line 6: I added the hyphen between cutting-edge cryptocurrency because when cutting edge is used as a noun it doesn't need a hyphen. When it is an adjective describing a noun (like in this case), it needs the hyphen.

Line 7: I capitalized the "Z" in Zerocoin technology and added a period at the end to keep it consistent with the rest of the bullet points.

Line 8: I capitalized "P" in Protocol as "Proof of Stake 3.0 Consensus Protocol" is one thing or could be written as "Proof of Stake 3.0 consensus protocol", but to keep it consistent would entail all capital or consensus protocol being lowercase. I also added a period after and before the next sentence with "Therefore," to make it a complete sentence, following the format of the 4th bullet point on line 11.

Line 9: I broke this up into two sentences, which I felt read easier and kept it in the same format as Line 11 and the previous bullet points I edited.

Line 13: "17.6M phore" was changed to 17.6M Phore(PHR) because later in the line you refer to PHR and not phore, not to mention Phore is always represented with a capital "P" because it is the name of the organization.

Line 41: I adjusted the grammar in here to make it into two sentences, which I believe should read easier.

*Side note: English can be written in many ways, I just felt that these changes help keep the README file more consistent. Feel free to reject or comment on the changes. Cheers!
# Conflicts:
#	README.md